// src/enums/TournamentAction.ts
export enum TournamentActions {
  FETCH_ALL_TOURNAMENT_START = 'FETCH_ALL_TOURNAMENT_START',
  FETCH_ALL_TOURNAMENT_SUCCESS = 'FETCH_ALL_TOURNAMENT_SUCCESS',
  FETCH_ALL_TOURNAMENT_ERROR = 'FETCH_ALL_TOURNAMENT_ERROR',
  SEARCH_TOURNAMENT_START = 'SEARCH_TOURNAMENT_START',
  SEARCH_TOURNAMENT_SUCCESS = 'SEARCH_TOURNAMENT_SUCCESS',
  SEARCH_TOURNAMENT_FAIL = 'SEARCH_TOURNAMENT_FAIL',
  CREATE_TOURNAMENT = 'CREATE_TOURNAMENT',
  EDIT_TOURNAMENT = 'EDIT_TOURNAMENT',
  DELETE_TOURNAMENT = 'DELETE_TOURNAMENT'
}